# backend/docs/swagger.yaml
definitions:
...
  entity.TicketType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - TicketTypeFree
    - TicketTypePaid
...
  sponsor_dto.CreateSponsor:
    properties:
      contact_info:
        type: string
      contribution_amount:
        type: number
      event_name:
        type: string
      name:
        type: string
    type: object
  ticket_dto.CreateTicket:
    properties:
      event_name:
        type: string
      login:
        type: string
      price:
        type: number
      ticket_type:
        $ref: '#/definitions/entity.TicketType'
    type: object
...
info:
  contact: {}
paths:
...
  /sponsors/{name}:
    get:
      description: Fetches the sponsor details by name
      parameters:
      - description: Sponsor name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get sponsor by name
      tags:
      - sponsors
    put:
      consumes:
      - application/json
      description: Updates the sponsor information such as name, contact, and event
        association
      parameters:
      - description: Sponsor name
        in: path
        name: name
        required: true
        type: string
      - description: Sponsor updates
        in: body
        name: sponsor
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a sponsor
      tags:
      - sponsors
  /sponsors/{sponsorName}:
    delete:
      description: Deletes the sponsor specified by the name
      parameters:
      - description: Sponsor name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a sponsor
      tags:
      - sponsors
  /sponsors/all:
    get:
      description: Fetches a list of all sponsors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all sponsors
      tags:
      - sponsors
  /sponsors/create:
    post:
      consumes:
      - application/json
      description: Creates a new sponsor with the provided details
      parameters:
      - description: Sponsor details
        in: body
        name: sponsor
        required: true
        schema:
          $ref: '#/definitions/sponsor_dto.CreateSponsor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a sponsor
      tags:
      - sponsors
  /tickets/{login}/{event_name}:
    delete:
      description: Удаляет билет по логину пользователя и названию события
      parameters:
      - description: Логин пользователя
        in: path
        name: userLogin
        required: true
        type: string
      - description: Название события
        in: path
        name: eventName
        required: true
        type: string
      responses:
        "200":
          description: Успешное удаление
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Удалить билет
      tags:
      - Tickets
    get:
      description: Возвращает билет по логину пользователя и названию события
      parameters:
      - description: Логин пользователя
        in: path
        name: userLogin
        required: true
        type: string
      - description: Название события
        in: path
        name: eventName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о билете
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить билет
      tags:
      - Tickets
    put:
      consumes:
      - application/json
      description: Обновляет данные билета
      parameters:
      - description: Логин пользователя
        in: path
        name: userLogin
        required: true
        type: string
      - description: Название события
        in: path
        name: eventName
        required: true
        type: string
      - description: Обновляемые данные
        in: body
        name: updatedTicket
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Обновить билет
      tags:
      - Tickets
  /tickets/all:
    get:
      description: Возвращает все билеты
      produces:
      - application/json
      responses:
        "200":
          description: Список билетов
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Список билетов
      tags:
      - Tickets
  /tickets/create:
    post:
      consumes:
      - application/json
      description: Создаёт новый билет для события
      parameters:
      - description: Данные билета
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/ticket_dto.CreateTicket'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание билета
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Создать билет
      tags:
      - Tickets
...
swagger: "2.0"